Combine

+++ remove_classes.py
@@ -1,6 +1,6 @@
 import json
 import argparse
-
+import os
 baltimore_ai_class_dict={ 
             0:"plastic_bag",
             1:"plastic_bottle",
@@ -19,7 +19,7 @@ baltimore_ai_class_dict={
             14:"misc"
             }
 
-def remove_classes_from_annotations(annotations_path, classes_to_remove, output_file,class_dict)->None:
+def remove_classes_from_annotations(annotations_path, classes_to_remove, output_file,class_dict = baltimore_ai_class_dict)->None:
     
     """
     Remove classes from annotations file
@@ -38,19 +38,23 @@ def remove_classes_from_annotations(annotations_path, classes_to_remove, output_
     annotations['annotations'] = [c for c in annotations['annotations'] if c['category_id'] not in classes_to_remove]
     annotations['categories'] = [c for c in annotations['categories'] if c['id'] not in classes_to_remove]
 
-
-    with open(output_file, 'w') as f:
-        json.dump(annotations, f)
+    if os.path.exists(output_file):
+        with open(output_file, 'w') as f:
+            json.dump(annotations, f)
+    else:
+        os.makedirs(os.path.dirname(output_file), exist_ok=True)
+        with open(output_file, 'w') as f:
+            json.dump(annotations, f)
         
 parser = argparse.ArgumentParser(description='Remove classes from annotations file')
 parser.add_argument('--annotations_path', type=str, help='Path to annotations file')
 parser.add_argument('--classes_to_remove', nargs='+',type=int, help='List of classes to remove seperated by space i.e. 0 1 2')
-parser.add_argument('--output_file', type=str, help='Path to output file')
-args = parser.parse_args()
+parser.add_argument('--output_file',default='coco_classes_removed.json', type=str, help='Path to output file')
 
-classes_to_remove = args.classes_to_remove   
-annotations_path = args.annotations_path
-output_file = args.output_file
+
+
 class_dict = baltimore_ai_class_dict
 
-remove_classes_from_annotations(annotations_path, classes_to_remove, output_file,class_dict)
+if __name__ == '__main__':
+    args = parser.parse_args()
+    remove_classes_from_annotations(args.annotations_path, args.classes_to_remove, args.output_file,class_dict)
